+++ src/jitterentropy-base.c
@@ -526,25 +526,12 @@ static struct rand_data
 		goto err;
 
 	/*
-	 * Initialize the hash state for the DRBG: FIPS is initialized as
-	 * SHA3-512 XDRBG-like conditioning component, otherwise as XDRBG-256.
-	 *
-	 * Ideally, the XDRBG-256 is always used, but as of now, it is no vetted
-	 * conditioning component and thus cannot be considered to provide full
-	 * entropy according to SP800-90B. Therefore, use a SHA3-512 XDRBG-like
-	 * conditioning component that is considered a vetted conditioning
-	 * component.
-	 *
-	 * When SP800-90A is updated to add the XDRBG-256 and thus SP800-90B
-	 * allows it as a vetted conditioning component, the SHA3-512 XDRBG-like
-	 * implementation is dropped.
+	 * Initialize the hash state for the XDRBG
 	 */
-	if ((flags & JENT_FORCE_FIPS) || jent_fips_enabled()) {
-		jent_sha3_512_init(entropy_collector->hash_state);
+	jent_shake256_init(entropy_collector->hash_state);
+
+	if ((flags & JENT_FORCE_FIPS) || jent_fips_enabled())
 		entropy_collector->fips_enabled = 1;
-	} else {
-		jent_shake256_init(entropy_collector->hash_state);
-	}
 
 	/* Set the oversampling rate */
 	entropy_collector->osr = osr;
@@ -590,8 +577,8 @@ static struct rand_data
 	/*
 	 * Assure, that we always have 512 bit entropy in our hash state
 	 * before outputting a block by adding at least 256 bit before first
-	 * usage. 256 bit (SHA3-512 XDRBG-like) or 512 bit (XDRBG-256) are
-	 * always transferred to the next state after the generation completes.
+	 * usage. 512 bit (XDRBG-256) are always transferred to the next state
+	 * after the generation completes.
 	 *
 	 * For NTG.1: already perform the startup stages here.
 	 */
@@ -801,14 +788,14 @@ out:
 	return ret;
 }
 
-static inline int jent_entropy_init_common_pre(unsigned int flags)
+static inline int jent_entropy_init_common_pre(void)
 {
 	int ret;
 
 	jent_notime_block_switch();
 	jent_health_cb_block_switch();
 
-	if (jent_sha3_tester((flags & JENT_FORCE_FIPS) || jent_fips_enabled()))
+	if (jent_sha3_tester())
 		return EHASH;
 
 	ret = jent_gcd_selftest();
@@ -830,7 +817,7 @@ static inline int jent_entropy_init_common_post(int ret)
 JENT_PRIVATE_STATIC
 int jent_entropy_init(void)
 {
-	int ret = jent_entropy_init_common_pre(0);
+	int ret = jent_entropy_init_common_pre();
 
 	if (ret)
 		return ret;
@@ -848,7 +835,7 @@ int jent_entropy_init(void)
 JENT_PRIVATE_STATIC
 int jent_entropy_init_ex(unsigned int osr, unsigned int flags)
 {
-	int ret = jent_entropy_init_common_pre(flags);
+	int ret = jent_entropy_init_common_pre();
 
 	if (ret)
 		return ret;